monitor/pip_requirements.txt
redis>=2.10.3
pyserial>=2.6
numpy>=1.8.2
